# api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bahdcoder - More-recipes",
    "description": "More-Recipes provides a platform for users to share the awesome and exciting  recipe ideas they have invented or learnt.  Suppose a user comes up with a recipe,  he/she can post it on More-Recipes and  get feedback in form of reviews and votes from other users who explore that recipe. Users can also keep a list of their favorite recipes on the application.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://bahdcoder-more-recipes-swagger.herokuapp.com/api/v1",
      "variables": {}
    }
  ],
  "paths": {
...
    "/users/settings": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Update user settings",
        "description": "",
        "operationId": "UpdateUserSettings",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImZyYW50ei5iYWhkY29kZXIyQGdtYWlsLmNvbSIsImlhdCI6MTUxODI5OTA3M30.8p6zeDBNsN1iNFhH1xh9iBpsY5QKMCHfYx6tyltdXjM"
            },
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImZyYW50ei5iYWhkY29kZXIyQGdtYWlsLmNvbSIsImlhdCI6MTUxODI5OTA3M30.8p6zeDBNsN1iNFhH1xh9iBpsY5QKMCHfYx6tyltdXjM"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "application/x-www-form-urlencoded"
            },
            "example": "application/x-www-form-urlencoded"
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "reviewEmails",
                  "favoriteModifiedEmail"
                ],
                "type": "object",
                "properties": {
                  "reviewEmails": {
                    "type": "string",
                    "example": "1"
                  },
                  "favoriteModifiedEmail": {
                    "type": "string",
                    "example": "0"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
...
  }
}